KAFKA-6696 Trogdor should support destroying tasks
Implement destroying tasks and workers.  This means erasing all record
of them on the Coordinator and the Agent.

Workers should be identified by unique 64-bit worker IDs, rather than by
the names of the tasks they are implementing.  This ensures that when a
task is destroyed and re-created with the same task ID, the old workers
will be not be treated as part of the new task instance.

Fix some return results from RPCs.  In some cases RPCs were returning
values that were never used.  Attempting to re-create the same task ID
with different arguments should fail.  Add RequestConflictException to
represent HTTP error code 409 (CONFLICT) for this scenario.

If only one worker in a task stops, don't stop all the other workers for
that task, unless the worker that stopped had an error.