Improve fencing in ZkBackingStore
Fence by using the controllerEpoch in ZkBackingStore.  This will  be
important when handling communication with KafkaController.  This
functionality is in AbstractEvent so it will be easy to reuse elsewhere
as well.  As a bonus, this allows us to avoid using cancellation to
cancel a bunch of events when the ZK session drops.  This is good since
cancellation is difficult to understand Another simplification is that
we can get rid of checkIsStartedAndActive because the check for the
correct controller epoch replaces the check to see if we are active.

Some other small cleanups: use accessors in AbstractEvent rather than
direct access, rename BackingStore#deactivateIfActive to
BackingStore#resign.