Add MetaLogManager
Add MetaLogManager, which manages the log which the kip-500 controller
relies on.  It handles things like claiming leadership, renouncing
leadership, applying records that have been committed, and so forth.

Eventually, we will have a MetaLogManager that hooks up to Raft.  For
now, we have a simple implementation based on a local log file which is
shared.