KAFKA-12334: Add the KIP-500 metadata shell
The Kafka Metadata shell is a new command which allows users to
interactively examine the metadata stored in a KIP-500 cluster.

It can read the metadata from the controllers directly, by connecting to
them, or from a metadata snapshot on disk.  In the former case, the
quorum voters must be specified by passing the --controllers flag; in
the latter case, the snapshot file should be specified via --snapshot.

The metadata tool works by replaying the log and storing the state into
in-memory nodes.  These nodes are presented in a fashion similar to
filesystem directories.